tests: Fix conversion warnings

+++ tests/tests_id_ringbuffer.cpp
@@ -32,7 +32,7 @@ TEST_CASE("u_template_historybuf")
 	}
 	SECTION("behavior with one")
 	{
-		int32_t zero_inner_index;
+		int64_t zero_inner_index;
 		REQUIRE((zero_inner_index = u_id_ringbuffer_push_back(buffer, 0)) >= 0);
 		CHECK_FALSE(u_id_ringbuffer_is_empty(buffer));
 		CHECK(1 == u_id_ringbuffer_get_size(buffer));
@@ -94,8 +94,8 @@ TEST_CASE("u_template_historybuf")
 
 	SECTION("behavior with two")
 	{
-		int32_t zero_inner_index;
-		int32_t one_inner_index;
+		int64_t zero_inner_index;
+		int64_t one_inner_index;
 		REQUIRE((zero_inner_index = u_id_ringbuffer_push_back(buffer, 0)) >= 0);
 		REQUIRE((one_inner_index = u_id_ringbuffer_push_back(buffer, 1)) >= 0);
 		REQUIRE(zero_inner_index != one_inner_index);
@@ -187,9 +187,9 @@ TEST_CASE("u_template_historybuf")
 
 	SECTION("algorithm behavior with 3")
 	{
-		int32_t zero_inner_index;
-		int32_t two_inner_index;
-		int32_t four_inner_index;
+		int64_t zero_inner_index;
+		int64_t two_inner_index;
+		int64_t four_inner_index;
 		REQUIRE((zero_inner_index = u_id_ringbuffer_push_back(buffer, 0)) >= 0);
 		REQUIRE((two_inner_index = u_id_ringbuffer_push_back(buffer, 2)) >= 0);
 		REQUIRE((four_inner_index = u_id_ringbuffer_push_back(buffer, 4)) >= 0);
